+++ modules/src/ics04_channel/handler/verify.rs
@@ -1,4 +1,4 @@
-use crate::ics02_client::state::ClientState;
+use crate::ics02_client::client_state::ClientState;
 use crate::ics02_client::{client_def::AnyClient, client_def::ClientDef};
 use crate::ics03_connection::connection::ConnectionEnd;
 use crate::ics04_channel::channel::ChannelEnd;